doc2db-final-version/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
openai>=1.12.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
Pillow>=10.2.0
pypdf>=4.0.0
openpyxl>=3.1.0
pymupdf>=1.24.0
